[chore/#44] dummy data 정의 위치 수정
